+++ .gitignore
@@ -2,3 +2,4 @@ env
 .DS_Store
 data_output.json
 .env
+archive
