+++ README.md
@@ -1,7 +1,13 @@
-cab-llm
+# C@B-GPT
+<img src="static/SS.png" style="height:500px"></img>
 
-reference : https://python.langchain.com/docs/tutorials/rag/
+Forget the medieval Courses @ Brown site! We bring you the means to tailor your own coursework!
+### Data
 
+We have scraped the Courses @ Brown website for relevant details of each course - we further took this forward by iterating over Google queries and parsing the most relevant links to each course
+
+
+### How to set your keys for inference 
 ```
 touch .env
 LANGSMITH_TRACING="true"
@@ -9,11 +15,10 @@ LANGSMITH_API_KEY="..."
 OPENAI_API_KEY=
 ```
 
+### Usage
 ```
-from dotenv import load_dotenv
-import os
-
-load_dotenv()  # Load .env file
-api_key = os.getenv("API_KEY")
-print(api_key)
+pip install -r requirements.txt
+python main.py
 ```
+### Future plans!
+Scale it for your nearest university!
