+++ requirements.txt
@@ -4,58 +4,39 @@ aiosignal==1.3.2
 annotated-types==0.7.0
 anyio==4.8.0
 appnope==0.1.4
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
 asttokens==3.0.0
-async-lru==2.0.4
 attrs==25.1.0
-babel==2.17.0
 beautifulsoup4==4.12.3
-bleach==6.2.0
+blinker==1.9.0
 certifi==2025.1.31
-cffi==1.17.1
 charset-normalizer==3.4.1
+click==8.1.8
 comm==0.2.2
 dataclasses-json==0.6.7
 debugpy==1.8.12
 decorator==5.1.1
-defusedxml==0.7.1
 distro==1.9.0
 executing==2.2.0
-fastjsonschema==2.21.1
-fqdn==1.5.1
+Flask==3.1.0
 frozenlist==1.5.0
+googlesearch-python==1.3.0
 h11==0.14.0
 httpcore==1.0.7
 httpx==0.28.1
 httpx-sse==0.4.0
 idna==3.10
+inexactsearch==1.0.2
 ipykernel==6.29.5
 ipython==8.32.0
-ipywidgets==8.1.5
-isoduration==20.11.0
+itsdangerous==2.2.0
 jedi==0.19.2
 Jinja2==3.1.5
 jiter==0.8.2
 jq==1.8.0
-json5==0.10.0
 jsonpatch==1.33
 jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2024.10.1
-jupyter==1.1.1
-jupyter-console==6.6.3
-jupyter-events==0.11.0
-jupyter-lsp==2.2.5
 jupyter_client==8.6.3
 jupyter_core==5.7.2
-jupyter_server==2.15.0
-jupyter_server_terminals==0.5.3
-jupyterlab==4.3.5
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.27.3
-jupyterlab_widgets==3.0.13
 langchain==0.3.17
 langchain-community==0.3.16
 langchain-core==0.3.33
@@ -68,70 +49,50 @@ langsmith==0.3.4
 MarkupSafe==3.0.2
 marshmallow==3.26.0
 matplotlib-inline==0.1.7
-mistune==3.1.1
 msgpack==1.1.0
 multidict==6.1.0
 mypy-extensions==1.0.0
-nbclient==0.10.2
-nbconvert==7.16.6
-nbformat==5.10.4
 nest-asyncio==1.6.0
-notebook==7.3.2
-notebook_shim==0.2.4
 numpy==1.26.4
 openai==1.61.0
 orjson==3.10.15
-overrides==7.7.0
 packaging==24.2
-pandocfilters==1.5.1
 parso==0.8.4
 pexpect==4.9.0
 platformdirs==4.3.6
-prometheus_client==0.21.1
 prompt_toolkit==3.0.50
 propcache==0.2.1
 psutil==6.1.1
 ptyprocess==0.7.0
 pure_eval==0.2.3
-pycparser==2.22
 pydantic==2.10.6
 pydantic-settings==2.7.1
 pydantic_core==2.27.2
 Pygments==2.19.1
 python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
-python-json-logger==3.2.1
 PyYAML==6.0.2
 pyzmq==26.2.1
-referencing==0.36.2
 regex==2024.11.6
 requests==2.32.3
 requests-toolbelt==1.0.0
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rpds-py==0.22.3
-Send2Trash==1.8.3
+silpa_common==0.3
 six==1.17.0
 sniffio==1.3.1
+soundex==1.1.3
 soupsieve==2.6
+spellchecker==0.4
 SQLAlchemy==2.0.37
 stack-data==0.6.3
 tenacity==9.0.0
-terminado==0.18.1
 tiktoken==0.8.0
-tinycss2==1.4.0
 tornado==6.4.2
 tqdm==4.67.1
 traitlets==5.14.3
-types-python-dateutil==2.9.0.20241206
 typing-inspect==0.9.0
 typing_extensions==4.12.2
-uri-template==1.3.0
 urllib3==2.3.0
 wcwidth==0.2.13
-webcolors==24.11.1
-webencodings==0.5.1
-websocket-client==1.8.0
-widgetsnbextension==4.0.13
+Werkzeug==3.1.3
 yarl==1.18.3
 zstandard==0.23.0
