--- scrapGoogle.py
@@ -1,5 +1,0 @@
-from googlesearch import search
-
-query = "CSCI 2240"
-for result in search(query):
-    print(result)
