--- script.py
@@ -1,69 +1,0 @@
-from flask import Flask, request, render_template, jsonify
-from dotenv import load_dotenv
-import os
-from langchain_openai import ChatOpenAI
-from langchain_openai import OpenAIEmbeddings
-from langchain_core.vectorstores import InMemoryVectorStore
-from langchain_community.document_loaders import TextLoader
-from langchain_text_splitters import RecursiveCharacterTextSplitter
-from langgraph.graph import START, StateGraph
-from typing_extensions import List, TypedDict
-from langchain import hub
-from langchain_core.documents import Document
-
-class State(TypedDict):
-    question: str
-    context: List[Document]
-    answer: str
-
-app = Flask(__name__)
-
-# Load environment variables
-load_dotenv()
-
-# Initialize LLM and vector store
-llm = ChatOpenAI(model="gpt-4o-mini")
-embeddings = OpenAIEmbeddings(model="text-embedding-3-large")
-vector_store = InMemoryVectorStore(embeddings)
-
-# Load and process text data
-path = "output.txt"  # Path to your .txt file
-loader = TextLoader(file_path=path)
-docs = loader.load()
-text_splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=20)
-all_splits = text_splitter.split_documents(docs)
-vector_store.add_documents(documents=all_splits)
-
-# Define the retrieval and generation functions
-def retrieve(state: State):
-    retrieved_docs = vector_store.similarity_search(state["question"], k=2)
-    return {"context": retrieved_docs}
-
-def generate(state: State):
-    docs_content = "\n\n".join(doc.page_content for doc in state["context"])
-    prompt = hub.pull("rlm/rag-prompt")
-    messages = prompt.invoke({"question": state["question"], "context": docs_content})
-    response = llm.invoke(messages)
-    return {"answer": response.content}
-
-# Construct the processing graph
-graph_builder = StateGraph(State)
-graph_builder.add_node("retrieve", retrieve)
-graph_builder.add_node("generate", generate)
-graph_builder.add_edge(START, "retrieve")
-graph_builder.add_edge("retrieve", "generate")
-graph = graph_builder.compile()
-
-@app.route("/")
-def home():
-    return render_template("index.html")
-
-@app.route("/ask", methods=["POST"])
-def ask():
-    data = request.json
-    question = data.get("question", "")
-    response = graph.invoke({"question": question})
-    return jsonify({"answer": response["answer"]})
-
-if __name__ == "__main__":
-    app.run(debug=True)
