+++ templates/index.html
@@ -14,7 +14,7 @@
     }
     body {
       font-family: 'Inter', sans-serif;
-      background: #f5f5f5;
+      background: url('static/background.jpg') no-repeat center center fixed;
       display: flex;
       align-items: center;
       justify-content: center;
